backend/requirements-dev.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
alembic
pydantic-settings
python-dotenv